backend/requirements.txt
Django==5.1.4
djangorestframework==3.16.0
django-cors-headers==4.7.0
django-filter==25.1
python-dotenv==1.0.0
Unidecode==1.3.8
black==24.10.0
pytest==8.3.4
pytest-django==4.9.0
factory-boy==3.3.1
faker==37.8.0

